Fix/front end (#13)
* Added loose index.html styling

* Updated TODO

* Various improvements to landing page design

* Added background image and adjusted size of text

* separated css into its own .css files by type

* Removed sizing.css since it was not necessary. Added comments to each .css file for better organizatoin

* Created images folder and moved bg.png into it

* Added text highlighting on the page

* Adjusted text highlighting

* Added barrel file to styles/ to be used in all components

* Set the app up to serve static content and started to break out dashboard.html into js, css, and html. Began debugging token issues as well

* Changed dashboard script.js to only set tokens based on url params. We will still need to account for when localStorage has params or there are no url params.

* Added check to make sure that we have localStorage tokens when we do not have param based tokens

* Changed tagline on index.html

* Added redesign for empty state in dashboard. Fixed bug in orphan fields not refreshing the page

* Started re-design of card component

* Updated TODO

* Cleanup on playlistController.post('/playlist')

* Completed phase 1 redesign of remix cards

* Changed naming of newly created playlists to be username remix

* Removed unnecessary log

* Changed instances of Playlist to Remix

* Added try catch to delete route to better handle errors

* Updated TODO

* Moved dequeuing log to be more accurate

* Added APIResponse interface to take advantage of refreshToken and accessToken showing on token refresh. Also updated middleware to show accessToken and refreshToken and started consuming in getUserPlaylists

* Updated todo

* Added APIResponse interface

* Added APIResponse to a few HTTP methods that are directly returned

* Added front end code in dashboard script.js to set tokens if refreshed by an api call, also changed history when refreshing to include songs that are only older than 6 days to avoid missing songs in refreshes when a user joins

* Added reroute to dashboard if param accessToken and refreshToken exist

* Began restructuring subscribe.html, added placeholder css and js files

* Primitive 401 failures added to server + UI, also stopped caching results in the browser

* Updated styling for history and subscribe page

* Adjusted filter to have a max of 2 songs from the same artist per user per playlist iteration

* removed bad import

* Removed unnecessary logs

* Additional log removal and incremented to 3 maxSongsPerArtistPerUser

* Added feature to prevent more than 3 songs from the same artist showing up for a given user in the playlist

* Made likedSongs a require property of SongsByUser and set equal to empty array where not being used

* Added a limit to getLikedSongsIfNecessary to ensure that we only make as many api calls as we need to satisfy songsPerUser rather than making as many API calls as we need to get the entire liked songs of a given user

+++ src/services/spotify/spotify.service.ts
@@ -23,30 +23,37 @@ export class SpotifyService {
       where: { accessToken },
       relations: ['memberPlaylists', 'ownedPlaylists'],
     });
-    return this.httpService.getUserPlaylists(accessToken).then(
-      (resp: AxiosResponse<SpotifyResponse<SpotifyPlaylist[]>>): PlaylistData => {
-        if (resp) {
-          const spotifyPlaylists: SpotifyPlaylist[] = resp.data.items;
-          if (user) {
-            const createdPlaylists = user?.ownedPlaylists?.map(x => x.playlistId) || [];
-            const memberPlaylists =
-              user?.memberPlaylists?.map(x => x.playlistId)?.filter(x => !createdPlaylists?.includes(x)) || [];
+    return this.httpService
+      .getUserPlaylists(accessToken)
+      .then(
+        (resp: AxiosResponse<SpotifyResponse<SpotifyPlaylist[]>>): PlaylistData => {
+          if (resp) {
+            const spotifyPlaylists: SpotifyPlaylist[] = resp.data.items;
+            if (user) {
+              const createdPlaylists = user?.ownedPlaylists?.map(x => x.playlistId) || [];
+              const memberPlaylists =
+                user?.memberPlaylists?.map(x => x.playlistId)?.filter(x => !createdPlaylists?.includes(x)) || [];
 
-            const orphanPlaylists = createdPlaylists.filter(x => !spotifyPlaylists.find(y => x === y.id));
-            const ownedPlaylists = spotifyPlaylists.filter(x => createdPlaylists?.includes(x.id));
-            const subscribedPlaylists = spotifyPlaylists.filter(x => memberPlaylists?.includes(x.id));
+              const orphanPlaylists = createdPlaylists.filter(x => !spotifyPlaylists.find(y => x === y.id));
+              const ownedPlaylists = spotifyPlaylists.filter(x => createdPlaylists?.includes(x.id));
+              const subscribedPlaylists = spotifyPlaylists.filter(x => memberPlaylists?.includes(x.id));
 
-            return {
-              ownedPlaylists,
-              orphanPlaylists,
-              subscribedPlaylists,
-            };
+              return {
+                ownedPlaylists,
+                orphanPlaylists,
+                subscribedPlaylists,
+                refreshToken: resp?.data?.refreshToken,
+                accessToken: resp?.data?.accessToken,
+              };
+            }
+            return { ownedPlaylists: [], orphanPlaylists: [], subscribedPlaylists: [] };
           }
           return { ownedPlaylists: [], orphanPlaylists: [], subscribedPlaylists: [] };
-        }
-        return { ownedPlaylists: [], orphanPlaylists: [], subscribedPlaylists: [] };
-      },
-    );
+        },
+      )
+      .catch(e => {
+        throw e;
+      });
   }
 
   async createUserPlaylist(accessToken: string): Promise<void> {
@@ -65,7 +72,10 @@ export class SpotifyService {
     const playlistsOwnedAndToBeDeleted: Playlist[] = ownedPlaylists.filter(x =>
       playlists.some(y => x.playlistId === y),
     );
-    return this.userService.deletePlaylist(playlistsOwnedAndToBeDeleted);
+    if (playlistsOwnedAndToBeDeleted.length) {
+      return this.userService.deletePlaylist(playlistsOwnedAndToBeDeleted);
+    }
+    throw new Error(`Unable to find playlist with id ${playlists}`);
   }
 
   unsubscribeFromPlaylist(_accessToken: string, _playlistId: string): void {
@@ -95,14 +105,29 @@ export class SpotifyService {
 
   async getTopSongs(members: User[], history: string[]): Promise<SongsByUser[]> {
     const historyIds = history;
+    const maxSongsPerArtistPerUser = 3;
     return members?.length
       ? await Promise.all(
           members.map(async (member: User) =>
             this.httpService.getTopSongsByUser(member).then(x => {
-              x.topSongs = x.topSongs.filter(song => !historyIds.includes(song.uri));
+              const seenArtists: Record<string, number> = {};
+
+              x.topSongs = x.topSongs.filter(song => {
+                let shouldSongBeIgnored = false;
+                // This ensures that we only allow a given maxSongsPerArtistPerUser so that we do not get entire albums from one person.
+                song.artists.forEach(artist => {
+                  seenArtists[artist.id] = seenArtists[artist.id] ? ++seenArtists[artist.id] : 1;
+                  if (seenArtists[artist.id] > maxSongsPerArtistPerUser) {
+                    shouldSongBeIgnored = true;
+                  }
+                });
+                return !historyIds.includes(song.uri) && !shouldSongBeIgnored;
+              });
+
               x.topSongs.forEach(song => {
                 historyIds.push(song.uri);
               });
+
               return x;
             }),
           ),
@@ -116,6 +141,7 @@ export class SpotifyService {
     );
   }
 
+  // may need advanced filtering here to filter out songs per user.
   getLikedSongsIfNecessary(
     songsByUser: SongsByUser[],
     songsPerUser: number,
@@ -123,7 +149,7 @@ export class SpotifyService {
   ): Promise<SongsByUser>[] {
     const historyIds = history;
     return songsByUser.map(async (x: SongsByUser) => {
-      if (x.topSongs.length < songsPerUser) {
+      if (x.topSongs.length < songsPerUser && x.likedSongs.length < songsPerUser) {
         x.topSongs.forEach(song => historyIds.push(song.uri));
         const newSongsByUser: SongsByUser = Object.assign({}, x);
         return this.httpService.getLikedSongsByUser(x.user).then(y => {
@@ -136,39 +162,26 @@ export class SpotifyService {
     });
   }
 
-  getRandomSongs(songs: SongWithUserData[], songsPerUser: number, playlistSongs: SongWithUserData[], count = 0): void {
-    const randomNumbers: Record<number, boolean> = {};
-    while (count < songsPerUser) {
-      const randomNumber = Math.floor(Math.random() * songs.length);
-      if (!randomNumbers[randomNumber]) {
-        randomNumbers[randomNumber] = true;
-        playlistSongs.push(songs[randomNumber]);
-        count += 1;
-      }
-    }
-  }
-
   generatePlaylist(music: SongsByUser[], songsPerUser: number): SongWithUserData[] {
-    console.log(songsPerUser);
     const playlistSongs: SongWithUserData[] = [];
     music.forEach((songsByUser: SongsByUser) => {
       const { topSongs, likedSongs } = songsByUser;
 
       const hasTopSongs = topSongs.length !== 0;
       const hasEnoughTopSongs = topSongs.length >= songsPerUser;
-      const hasEnoughLikedSongs = !!likedSongs && likedSongs.length >= songsPerUser;
+      const hasLikedSongs = likedSongs && likedSongs.length !== 0;
+      const hasEnoughLikedSongs = hasLikedSongs && likedSongs.length >= songsPerUser;
       const hasEnoughTopSongsAndLikedSongs =
-        hasTopSongs && !!likedSongs && topSongs.length + likedSongs.length >= songsPerUser;
+        hasTopSongs && hasLikedSongs && topSongs.length + likedSongs.length >= songsPerUser;
 
-      if (!hasTopSongs && !!likedSongs && hasEnoughLikedSongs) {
-        return this.getRandomSongs(likedSongs, songsPerUser, playlistSongs);
+      if (!hasTopSongs && hasLikedSongs && hasEnoughLikedSongs) {
+        playlistSongs.concat(likedSongs.slice(0, songsPerUser - 1));
       } else if (hasEnoughTopSongs) {
-        for (let i = 0; i < songsPerUser; i++) {
-          playlistSongs.push(topSongs[i]);
-        }
-      } else if (!hasEnoughTopSongs && !!likedSongs && hasEnoughTopSongsAndLikedSongs) {
+        playlistSongs.concat(topSongs.slice(0, songsPerUser - 1));
+      } else if (!hasEnoughTopSongs && hasEnoughTopSongsAndLikedSongs) {
         topSongs.forEach(song => playlistSongs.push(song));
-        this.getRandomSongs(likedSongs, songsPerUser, playlistSongs, topSongs.length);
+        playlistSongs.concat(topSongs);
+        playlistSongs.concat(likedSongs.slice(0, playlistSongs.length - songsPerUser));
       }
     });
 
@@ -176,7 +189,7 @@ export class SpotifyService {
   }
 
   async getAllMusic(members: User[], songsPerUser: number, history: Song[]): Promise<SongWithUserData[]> {
-    const historyIds: string[] = this.filterByHours(history, 'createdAt', 12).map(x => x.spotifyUrl);
+    const historyIds: string[] = this.filterByHours(history, 'createdAt', 144).map(x => x.spotifyUrl);
 
     let music: SongsByUser[] = await this.getTopSongs(members, historyIds);
 
@@ -203,7 +216,6 @@ export class SpotifyService {
       const { members, history, owner } = playlist;
 
       const songsPerUser = this.getNumberOfItemsPerUser(members.length);
-      console.log(songsPerUser);
       const music: SongWithUserData[] = await this.getAllMusic(members, songsPerUser, history);
       const orderedPlaylist: SongWithUserData[] = this.roundRobinSort(music);
       // Get all songs from the playlist.
@@ -220,7 +232,11 @@ export class SpotifyService {
     return undefined;
   }
 
-  removeAllPlaylistTracks(playlistId: string, accessToken: string, playlistTracks: SpotifyPlaylistItemInfo[]) {
+  removeAllPlaylistTracks(
+    playlistId: string,
+    accessToken: string,
+    playlistTracks: SpotifyPlaylistItemInfo[],
+  ): Promise<any> {
     const calls = [];
     if (playlistTracks.length > 100) {
       const numberOfCalls = Math.ceil(playlistTracks.length / 100);
@@ -237,8 +253,6 @@ export class SpotifyService {
     return Promise.all(calls);
   }
 
-  // Note: This sort is a "best effort" to maintain order within the playlist.
-  // If a failure occurs or one request completes sooner than another the order will not be maintained.
   roundRobinSort(arr: SongWithUserData[]): SongWithUserData[] {
     const allSongs = arr;
     let sortedArr: SongWithUserData[] = [];
